move handlers and storage to internal
