+++ cmd/server/main.go
@@ -9,7 +9,7 @@ import (
 
 	_ "net/http/pprof" // подключаем пакет pprof
 
-	"github.com/kmx0/devops/cmd/server/handlers"
+	"github.com/kmx0/devops/internal/handlers"
 	"github.com/kmx0/devops/internal/config"
 	"github.com/sirupsen/logrus"
 )
