+++ internal/handlers/compress.go
