+++ internal/handlers/handlers.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/gin-contrib/pprof"
 	"github.com/gin-gonic/gin"
-	"github.com/kmx0/devops/cmd/server/storage"
+	"github.com/kmx0/devops/internal/storage"
 	"github.com/kmx0/devops/internal/config"
 	"github.com/kmx0/devops/internal/crypto"
 	"github.com/kmx0/devops/internal/repositories"
